Time: 49 ms (69.33%), Space: 149 MB (33.47%) - LeetHub

+++ 3974-maximum-total-sum-of-k-selected-elements/3974-maximum-total-sum-of-k-selected-elements.java
@@ -5,7 +5,7 @@ class Solution {
         long res = 0;
         Arrays.sort(nums);
         for(int i = n-1;i >= (n-k); i--){
-            int num = nums[i];
+            long num = nums[i];
             res += num*mul;
             mul--;
             if(mul == 0) mul = 1;
